fix(user-library): apply code quality fixes to UserHandRepository
- Fix 1: Replace DateTime.UtcNow with TimeProvider.GetUtcNow().UtcDateTime
- Fix 2: Remove SaveChangesAsync from repository (delegate to IUnitOfWork in CQRS handlers, matching WishlistRepository pattern)
- Fix 3: Extract UserHandSlotData record to its own file Domain/Repositories/UserHandSlotData.cs
