+++ apps/api/src/Api/BoundedContexts/UserLibrary/Domain/Repositories/IUserHandRepository.cs
@@ -6,13 +6,3 @@ public interface IUserHandRepository
     Task UpsertSlotAsync(Guid userId, string slotType, Guid entityId, string entityType, string? entityLabel, string? entityImageUrl, CancellationToken ct = default);
     Task ClearSlotAsync(Guid userId, string slotType, CancellationToken ct = default);
 }
-
-/// <summary>Simple data transfer record used within the domain layer.</summary>
-public record UserHandSlotData(
-    string SlotType,
-    Guid? EntityId,
-    string? EntityType,
-    string? EntityLabel,
-    string? EntityImageUrl,
-    DateTime? PinnedAt
-);
