+++ apps/api/src/Api/BoundedContexts/UserLibrary/Domain/Repositories/UserHandSlotData.cs
@@ -1,0 +1,11 @@
+namespace Api.BoundedContexts.UserLibrary.Domain.Repositories;
+
+/// <summary>Simple data transfer record used within the domain layer.</summary>
+public record UserHandSlotData(
+    string SlotType,
+    Guid? EntityId,
+    string? EntityType,
+    string? EntityLabel,
+    string? EntityImageUrl,
+    DateTime? PinnedAt
+);
